% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix-package.R
\docType{package}
\name{admix-package}
\alias{admix}
\alias{admix-package}
\title{admix: Package admix for admixture (aka contamination) models}
\description{
This package implements several methods to estimate the
    unknown quantities related to two-component admixture models,
    depending on the assumptions made on the unknown component density. In
    practice, one can estimate both the mixture weight and the unknown
    component density in a wide variety of frameworks. On top of that,
    hypothesis tests can be performed in one and two-samples contexts to
    test the unknown component density. Finally, clustering of unknown
    mixture components is also feasible in a K-samples setting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/XavierMilhaud/admix}
  \item Report bugs at \url{https://github.com/XavierMilhaud/admix/issues}
}

}
\author{
\strong{Maintainer}: Xavier Milhaud \email{xavier.milhaud.research@gmail.com} [contributor]

}
\keyword{internal}
