% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_estim.R
\name{admix_estim}
\alias{admix_estim}
\title{Estimate the unknown weight in an admixture model}
\usage{
admix_estim(samples, admixMod, est_method = c("PS", "BVdk", "IBM"), ...)
}
\arguments{
\item{samples}{A list of the K (K>0) samples to be studied, all following admixture distributions.}

\item{admixMod}{A list of objects of class \link[admix]{admix_model}, with information about known distributions and known parameters.}

\item{est_method}{The estimation method to be applied. Can be one of 'BVdk' (Bordes and Vandekerkhove estimator), 'PS' (Patra and Sen
estimator), or 'IBM' (Inversion Best-Matching approach) in the continuous case (continuous random variable). Only 'IBM' for
discrete random variables. The same estimation method is performed on each sample if several samples are provided.}

\item{...}{Optional arguments to \link[admix]{estim_PS}, \link[admix]{estim_BVdk} or \link[admix]{estim_IBM} depending on the
choice made by the user for the estimation method.}
}
\value{
An object of class \code{estim_BVdk}, \code{estim_PS} or \code{estim_IBM} (that inherits from class \link[admix]{admix_estim}),
with two attributes, 'class' and 'names'. The latter contains three elements, among which 'estim_objects' that lists for each
sample under study all the information of the estimation procedure.
}
\description{
Estimate the unknown component weight (and possibly a location shift parameter
in case of a symmetric unknown component density), using different estimation
techniques. We recall that the \eqn{i}-th admixture model has probability density function
\eqn{\ell_i} such that:
\deqn{
  \ell_i = p_i f_i + (1 - p_i) g_i,
}
where \eqn{g_i} is the known component density. The unknown quantities
\eqn{p_i} and \eqn{f_i} then have to be estimated.
}
\details{
For further details on the different estimation techniques, see references below on i) Patra and Sen estimator ;
ii) Bordes and Vandekerkhove estimator ; iii) Inversion Best-Matching approach. Important note: estimation by 'IBM'
requires at least two samples at hand, and provides unbiased estimators only if the distributions of unknown components
are equal (meaning that it requires to perform previously this test between the pairs of samples, see \link[admix]{admix_test}).
When selecting 'BVdk' estimation method, the initialization parameters for the optimization process have been arbitrarily
set. The mixing proportion is fixed to 0.5. For the localization parameter, it is based on taking the first moment
in the model (\eqn{\ell(x) = p f(x-\mu) + (1 - p) g(x)}), and isolating \eqn{\mu} by an inversion formula.
}
\examples{
## Simulate mixture data:
mixt1 <- twoComp_mixt(n = 300, weight = 0.7,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = 0, "sd" = 1)))
mixt2 <- twoComp_mixt(n = 250, weight = 0.85,
                      comp.dist = list("norm", "exp"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("rate" = 1)))
mixt3 <- twoComp_mixt(n = 500, weight = 0.5,
                      comp.dist = list("pois", "pois"),
                      comp.param = list(list("lambda" = 2),
                                        list("lambda" = 7)))
mixt4 <- twoComp_mixt(n = 1500, weight = 0.2, comp.dist = list("multinom", "multinom"),
                      comp.param = list(list("size"=1, "prob" = c(0.8,0.1,0.1)),
                                   list("size"=1, "prob" = c(0.1,0.2,0.7))))
data1 <- get_mixture_data(mixt1)
data2 <- get_mixture_data(mixt2)
data3 <- get_mixture_data(mixt3)
data4 <- get_mixture_data(mixt4)
## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])
admixMod3 <- admix_model(knownComp_dist = mixt3$comp.dist[[2]],
                         knownComp_param = mixt3$comp.param[[2]])
admixMod4 <- admix_model(knownComp_dist = mixt4$comp.dist[[2]],
                         knownComp_param = mixt4$comp.param[[2]])
# Estimation by different methods:
admix_estim(samples = list(data1), admixMod = list(admixMod1), est_method = "BVdk")
admix_estim(samples = list(data1, data2, data3, data4),
            admixMod = list(admixMod1, admixMod2, admixMod3, admixMod4), est_method = "PS")
admix_estim(samples = list(data1,data2), admixMod = list(admixMod1,admixMod2), est_method = "IBM")

}
\references{
\insertRef{PatraSen2016}{admix}
\insertRef{BordesDelmasVandekerkhove2006}{admix}
\insertRef{BordesVandekerkhove2010}{admix}
\insertRef{MilhaudPommeretSalhiVandekerkhove2024a}{admix}
}
\seealso{
\code{\link[=get_mixing_weights]{get_mixing_weights()}} to access the estimated mixing weight(s), \code{\link[=get_known_component]{get_known_component()}} to access the known component(s), \code{\link[=print.admix_estim]{print.admix_estim()}} for a brief description of the results,
and \code{\link[=summary.admix_estim]{summary.admix_estim()}} for an overview of the estimation process. More precisely, 1) the number of samples under study;
2) the information about the known mixture components (distributions and parameters); 3) the sizes of the samples;
4) the chosen estimation technique (one of 'BVdk', 'PS' or 'IBM'); 5) the estimated mixing proportions (weights of the
unknown component distributions in the mixture model). In case of 'BVdk' estimation, one additional attribute corresponding
to the estimated location shift parameter is included.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
