% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGalaxies.R
\docType{data}
\name{allGalaxies}
\alias{allGalaxies}
\title{Four galaxies measurements of heliocentric velocities (Carina, Sextans, Sculptor, Fornax)}
\format{
A data frame with ... rows and .. variables:
\describe{
\item{Target}{The machine}
\item{HV}{The value of heliocentric velocity}
\item{Name}{The name of the galaxy}
}
}
\source{
\url{https://iopscience.iop.org/article/10.1088/0004-6256/137/2/3100}
}
\usage{
allGalaxies
}
\description{
Four galaxies measurements of heliocentric velocities (Carina, Sextans, Sculptor, Fornax)
}
\keyword{datasets}
