% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distribution_type}
\alias{distribution_type}
\title{Determine the type of distribution under consideration}
\usage{
distribution_type(dist)
}
\arguments{
\item{dist}{A single string naming the distribution under consideration.}
}
\value{
The type of distribution under study (continuous, discrete, or multivariate).
}
\description{
Determine the type of distribution under consideration
}
\examples{
distribution_type("norm")
distribution_type("pois")
distribution_type("multinom")
distribution_type("weibull")

}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
