% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_cluster_members.admix_cluster}
\alias{get_cluster_members.admix_cluster}
\title{Extractor for object of class 'admix_cluster'}
\usage{
\method{get_cluster_members}{admix_cluster}(x)
}
\arguments{
\item{x}{An object of class 'admix_cluster'.}
}
\description{
Extract the clusters that were discovered among K samples, where belonging to
one given cluster means having equal unknown component distributions.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
