% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_cluster.R
\name{get_cluster_sizes}
\alias{get_cluster_sizes}
\title{Extractor for object of class 'admix_cluster'}
\usage{
get_cluster_sizes(x)
}
\arguments{
\item{x}{An object of class 'admix_cluster'.}
}
\description{
Provide the number of samples in each cluster.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
