% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_discrepancy_matrix.admix_cluster}
\alias{get_discrepancy_matrix.admix_cluster}
\title{Extractor for object of class 'admix_cluster'}
\usage{
\method{get_discrepancy_matrix}{admix_cluster}(x)
}
\arguments{
\item{x}{An object of class 'admix_cluster'.}
}
\description{
Provide the matrix storing discrepancies using Inversion-Best Matching
approach between all couples among the K samples under study.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
