% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_distribution_parameters}
\alias{get_distribution_parameters}
\title{Check the validity of the specified distributions}
\usage{
get_distribution_parameters(dist)
}
\arguments{
\item{dist}{A single string naming the distribution under consideration.}
}
\value{
A vector composed of the names of the parameters of the distribution.
}
\description{
Check the validity of the specified distributions
}
\examples{
get_distribution_parameters("norm")
get_distribution_parameters("gamma")
get_distribution_parameters("weibull")

}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
