% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_known_component.admix_estim}
\alias{get_known_component.admix_estim}
\title{Extractor for objects of class 'admix_estim'}
\usage{
\method{get_known_component}{admix_estim}(x)
}
\arguments{
\item{x}{An object of class 'admix_estim'.}
}
\description{
Get the known component(s) of admixture model(s) considered for estimation.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
