% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_known_component.htest}
\alias{get_known_component.htest}
\title{Extractor for objects of class 'htest'}
\usage{
\method{get_known_component}{htest}(x)
}
\arguments{
\item{x}{An object of class 'htest'.}
}
\description{
Get the known component(s) of admixture model(s) considered for testing
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
