% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_mixing_weights.htest}
\alias{get_mixing_weights.htest}
\title{Extractor for objects of class 'htest'}
\usage{
\method{get_mixing_weights}{htest}(x)
}
\arguments{
\item{x}{An object of class 'htest'.}
}
\description{
Get the estimated unknown mixing proportion(s) related to the weight(s) of
the unknown component distribution(s) of the admixture model(s).
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
