% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_estim.R
\name{get_mixing_weights}
\alias{get_mixing_weights}
\title{Extractor for object of class 'admix_estim'}
\usage{
get_mixing_weights(x)
}
\arguments{
\item{x}{An object of class 'admix_estim'.}
}
\description{
Get the estimated unknown mixing proportion related to the weight of
the unknown component distribution of the admixture model.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
