% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_tabulated_dist.admix_cluster}
\alias{get_tabulated_dist.admix_cluster}
\title{Extractor for object of class 'admix_cluster'}
\usage{
\method{get_tabulated_dist}{admix_cluster}(x)
}
\arguments{
\item{x}{An object of class 'admix_cluster'.}
}
\description{
Provide the list of tabulated distributions that allow to define
the extreme quantiles against which the test statistics are compared.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
