% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_tabulated_dist.IBM_test}
\alias{get_tabulated_dist.IBM_test}
\title{Extractor for object of class 'IBM_test'}
\usage{
\method{get_tabulated_dist}{IBM_test}(x)
}
\arguments{
\item{x}{An object of class 'IBM_test'.}
}
\description{
Provide the tabulated distribution that allows to define the extreme
quantile against which the test statistic is compared.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
