% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM_estimVarCov_gaussVect.R
\name{IBM_estimVarCov_gaussVect}
\alias{IBM_estimVarCov_gaussVect}
\title{Estimates the covariance matrix of the gaussian vector (IBM)}
\usage{
IBM_estimVarCov_gaussVect(x, y, IBMestim.obj, samples, admixMod)
}
\arguments{
\item{x}{Time point at which the 1st (related to the 1st parameter) underlying empirical process is looked through.}

\item{y}{Time point at which the 2nd (related to the 2nd parameter) underlying empirical process is looked through.}

\item{IBMestim.obj}{An object of class 'estim_IBM'.}

\item{samples}{(List) List of the two considered samples.}

\item{admixMod}{(List) List of objects of class 'admix_model', one for each sample.}
}
\value{
The estimated variance-covariance matrix of the gaussian vector Z = (hat(p1),(hat(p2),Dn(z)), at location '(x,y)'.
}
\description{
Nonparametric estimation of the covariance matrix of the gaussian vector at point 'z', considering the use of Inversion - Best
Matching (IBM) method to estimate the model parameters in two-sample admixture models.
Recall that the two admixture models have respective probability density functions (pdf) l1 and l2, such that:
l1 = p1*f1 + (1-p1)\emph{g1 and l2 = p2}f2 + (1-p2)*g2, where g1 and g2 are the known component densities.
Further information for the IBM approach are given in 'Details' below.
}
\examples{
\donttest{
## Simulate mixture data:
mixt1 <- twoComp_mixt(n = 2000, weight = 0.2,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = 2, "sd" = 3),
                                        list("mean" = -2, "sd" = 1)))
plot(mixt1, xlim = c(0,30), ylim = c(0,0.15))
data1 <- getmixtData(mixt1)
mixt2 <- twoComp_mixt(n = 1500, weight = 0.5,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = 2, "sd" = 3),
                                        list("mean" = 6, "sd" = 1)))
plot(mixt2, add.plot = TRUE, xlim = c(0,30), ylim = c(0,0.15), col = "blue")
data2 <- getmixtData(mixt2)

## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])

## Estimate the mixture weights of the two admixture models (provide only hat(theta)_n):
est <- estim_IBM(samples = list(data1,data2),
                 admixMod = list(admixMod1,admixMod2), n.integ = 1000)

IBM_estimVarCov_gaussVect(x = mean(data1), y = mean(data2), IBMestim.obj = est,
                          samples=list(data1,data2), admixMod = list(admixMod1,admixMod2))
}

}
\references{
\insertRef{MilhaudPommeretSalhiVandekerkhove2024a}{admix}
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
