% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_cdf.R
\name{kernel_cdf}
\alias{kernel_cdf}
\title{Kernel estimation}
\usage{
kernel_cdf(u, h)
}
\arguments{
\item{u}{the point at which the estimation is made.}

\item{h}{window of the kernel estimation.}
}
\value{
the estimated value of the cdf.
}
\description{
Functions to perform the estimation of cumulative distribution function (cdf) by kernel estimators
(with a non-gaussian kernel).
}
\examples{
kernel_cdf(0.4,0.5)

}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
