% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milkyWay.R
\docType{data}
\name{milkyWay}
\alias{milkyWay}
\title{Heliocentric velocity measured from the Milky Way.}
\format{
A data frame with ... rows and .. variables:
\describe{
\item{V1}{The heliocentric velocity of Milky way}
}
}
\source{
\url{https://www.aanda.org/articles/aa/full_html/2018/08/aa32905-18/aa32905-18.html}
}
\usage{
milkyWay
}
\description{
Heliocentric velocity measured from the Milky Way.
}
\keyword{datasets}
