% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoBasis_test_H0.R
\name{orthoBasis_test_H0}
\alias{orthoBasis_test_H0}
\title{Equality test of unknown components between two admixture models using polynomial basis expansions}
\usage{
orthoBasis_test_H0(
  data.X,
  data.Y,
  known.p = NULL,
  comp.dist = NULL,
  comp.param = NULL,
  known.coef = NULL,
  K = 3,
  nb.ssEch = 2,
  s = 0.49,
  var.explicit = FALSE,
  nb.echBoot = NULL,
  support = c("Real", "Integer", "Positive", "Bounded.continuous", "Bounded.discrete"),
  bounds.supp = NULL,
  est.method = c("BVdk", "PS"),
  uniformized.knownComp_data = NULL
)
}
\arguments{
\item{data.X}{First observed sample following mixture distribution given by l1.}

\item{data.Y}{Second observed sample following mixture distribution given by l2.}

\item{known.p}{(default to NULL) Numeric vector with two elements, respectively the component weight for the unknown component
in the first and in the second samples.}

\item{comp.dist}{A list with four elements corresponding to the component distributions (specified with R native names for these distributions)
involved in the two admixture models. The two first elements refer to the unknown and known components of the 1st admixture model,
and the last two ones to those of the second admixture model. If there are unknown elements, they must be specified as 'NULL' objects.
For instance, 'comp.dist' could be specified as follows: list(f1=NULL, g1='norm', f2=NULL, g2='norm').}

\item{comp.param}{A list with four elements corresponding to the parameters of the component distributions, each element being a list
itself. The names used in this list must correspond to the native R argument names for these distributions.
The two first elements refer to the parameters of unknown and known components of the 1st admixture model, and the last
two ones to those of the second admixture model. If there are unknown elements, they must be specified as 'NULL' objects.
For instance, 'comp.param' could be specified as follows: : list(f1=NULL, g1=list(mean=0,sd=1), f2=NULL, g2=list(mean=3,sd=1.1)).}

\item{known.coef}{Coefficients in the polynomial basis expansion, corresponding to the known component densities g1 and g2.}

\item{K}{Number of coefficients considered for the polynomial basis expansion.}

\item{nb.ssEch}{Number of subsamples created from the original data to decorrelate the estimation of the different parameters.}

\item{s}{Rate at which the normalization factor is set in the penalization rule for model selection (in ]0,1/2[), see 'Details'.}

\item{var.explicit}{Boolean that allows to choose between explicit assessment of the variance of the test statistic or not (FALSE=bootstrap),
FIXME : it seems that bootstrap procedure does not work in the context of admixtures.}

\item{nb.echBoot}{number of bootstrap samples if 'var.explicit' is set to FALSE.}

\item{support}{support of the densities under consideration, useful to choose the polynomial orthonormal basis.}

\item{bounds.supp}{(default to NULL) useful if support = 'bounded', a list of minimum and maximum bounds, specified as
following: list( list(min.f1,min.g1,min.f2,min.g2) , list(max.f1,max.g1,max.f2,max.g2) )}

\item{est.method}{Estimation method to get the component weights, either 'PS' (Patra and Sen estimation) or 'BVdk' (Bordes and Vendekerkhove estimation).}

\item{uniformized.knownComp_data}{(default to NULL) Only useful if 'est.method' has been set to 'PS', and for real-life applications
where the distribution of the known component of the admixture model is also unknown. In this
case, this known component is previously made uniformly(0,1)-distributed by applying the empirical
cumulative distribution of the known component function on the data. This means that
all 'comp.dist' and 'comp.param' must be set to NULL.}
}
\value{
A list with six elements containing: 1) the rejection decision; 2) the p-value of the test; 3) the test statistic; 4) the
variance-covariance matrix of the test statistic; 5) selected rank for testing, and 6) estimates of the two component weights.
}
\description{
Test the null hypothesis (H0: f1=f2) using the decomposition of unknown densities of the two admixture distributions in
an adequate orthonormal polynomial basis. Recall that we have two admixture models with respective probability density
functions (pdf) l1 = p1*f1 + (1-p1)\emph{g1 and l2 = p2}f2 + (1-p2)*g2, where g1 and g2 are the only known elements.
The admixture weights p1 and p2 thus have to be estimated. For further information on this method, see 'Details' below.
}
\details{
See the paper on HAL website: https://hal.archives-ouvertes.fr/hal-02491127
}
\examples{
\dontrun{
###### Using Bordes and Vandekerkhove estimation (valid if symetric unknown component densities).
#### First under the null hypothesis H0.
## Simulate data:
list.comp <- list(f1 = "norm", g1 = "norm",
                  f2 = "norm", g2 = "norm")
list.param <- list(f1 = c(mean = 1, sd = 1), g1 = c(mean = 4, sd = 1),
                   f2 = c(mean = 1, sd = 1), g2 = c(mean = 5, sd = 0.5))
sim.X <- rsimmix(n = 2500, unknownComp_weight=0.7, comp.dist = list(list.comp$f1,list.comp$g1),
                 comp.param = list(list.param$f1, list.param$g1))
sim.Y <- rsimmix(n = 3000, unknownComp_weight=0.5, comp.dist = list(list.comp$f2,list.comp$g2),
                 comp.param = list(list.param$f2, list.param$g2))
plot_admix(sim.X = sim.X[['mixt.data']], sim.Y = sim.Y[['mixt.data']], support = "continuous")
## Perform the hypothesis test in real-life conditions:
list.comp <- list(f1 = NULL, g1 = "norm",
                  f2 = NULL, g2 = "norm")
list.param <- list(f1 = NULL, g1 = c(mean = 4, sd = 1),
                   f2 = NULL, g2 = c(mean = 5, sd = 0.5))
test <- orthoBasis_test_H0(data.X = sim.X[['mixt.data']], data.Y = sim.Y[['mixt.data']],
             known.p=NULL, comp.dist = list.comp, comp.param = list.param, known.coef=NULL, K=3,
             nb.ssEch = 2, s = 0.49, var.explicit=TRUE, nb.echBoot=NULL, support = 'Real',
             bounds.supp = NULL, est.method = 'BVdk', uniformized.knownComp_data = NULL)
test$decision

#### Then under the alternative hypothesis H1.
## Simulate data:
list.comp <- list(f1 = "norm", g1 = "norm",
                  f2 = "norm", g2 = "norm")
list.param <- list(f1 = c(mean = 1, sd = 1), g1 = c(mean = 4, sd = 1),
                   f2 = c(mean = 2, sd = 0.8), g2 = c(mean = 5, sd = 0.5))
sim.X <- rsimmix(n = 2500, unknownComp_weight=0.7, comp.dist = list(list.comp$f1,list.comp$g1),
                 comp.param = list(list.param$f1, list.param$g1))
sim.Y <- rsimmix(n = 3000, unknownComp_weight=0.5, comp.dist = list(list.comp$f2,list.comp$g2),
                 comp.param = list(list.param$f2, list.param$g2))
plot_admix(sim.X = sim.X[['mixt.data']], sim.Y = sim.Y[['mixt.data']], support = "continuous")
## Perform the hypothesis test in real-life setting:
list.comp <- list(f1 = NULL, g1 = "norm",
                  f2 = NULL, g2 = "norm")
list.param <- list(f1 = NULL, g1 = c(mean = 4, sd = 1),
                   f2 = NULL, g2 = c(mean = 5, sd = 0.5))
test <- orthoBasis_test_H0(data.X = sim.X[['mixt.data']], data.Y = sim.Y[['mixt.data']],
            known.p=NULL, comp.dist = list.comp, comp.param = list.param, known.coef=NULL, K=3,
            nb.ssEch = 2, s = 1, var.explicit = TRUE, nb.echBoot =NULL, support = 'Real',
            bounds.supp = NULL, est.method = 'BVdk', uniformized.knownComp_data = NULL)
test$decision
test$p1
test$p2

###### Real-life data: using Patra and Sen estimation (although not n square-root consistent).
## Goal: compare heliocentric velocities of different satellites.
data(allGalaxies)
HV <- allGalaxies[ ,c('HV','Name')]
HVcar <- HV[which(HV$Name == 'Carina'), ]
HVcar <- HVcar[-which(is.na(HVcar)), ]
HVcar <- as.numeric(HVcar$HV)
HVsex <- HV[which(HV$Name == 'Sextans'), ]
HVsex <- HVsex[-which(is.na(HVsex)), ]
HVsex <- as.numeric(HVsex$HV)
## Retrieve heliocentric velocity of the Milky way:
data(milkyWay)
MW <- milkyWay$V1
plot(density(MW), main = "Milky Way", xlab = "", xlim = c(-100,300))
plot(density(HVcar), main = "Carina", xlab = "", xlim = c(-100,300))
plot(density(HVsex), main = "Sextans", xlab = "", xlim = c(-100,300))
## Extraction of coef related to the Milky way HV density in the orthonormal basis expansion:
## Milky way is not a mixture, but is the known component in coming admixture distributions:
donnees.voieLactee <- list(data.brute = MW, data.transform = NULL)
summary(donnees.voieLactee[['data.brute']])
moy.voieLactee <- mean(donnees.voieLactee[['data.brute']])
var.voieLactee <- var(donnees.voieLactee[['data.brute']])
## Calcul de la fonction de repartition empirique:
empiricalCDF.MW <- ecdf(donnees.voieLactee[['data.brute']])
plot(empiricalCDF.MW)
## Calcul des coefficients dans la decomposition dans la base orthonormale:
coefs.voieLactee <- orthoBasis_coef(data = donnees.voieLactee[['data.brute']], supp = 'Real',
                                    degree = 3, m = 3, other = NULL)
coefs.voieLactee <- sapply(coefs.voieLactee, mean)
## Test the unknown densities between Carina and Sextans:
donnees.X <- HVcar
donnees.Y <- HVsex
## Formating data: transformation to a mixture with one known uniform distribution:
dat.X <- list(data.brute = donnees.X, data.transform = NULL)
plot_admix(sim.X = dat.X[['data.brute']], sim.Y = NULL, user.bounds=NULL, support="continuous")
## Densite apres avoir rendu la composante connue uniforme:
dat.X[['data.transform']] <- empiricalCDF.MW(dat.X[['data.brute']])
mean(dat.X[['data.transform']])   # > 0.5 means that known component is on the left hand side
plot_admix(sim.X=dat.X[['data.transform']], sim.Y=NULL, user.bounds=NULL, support="continuous")
## Same for the second sample:
dat.Y <- list(data.brute = donnees.Y, data.transform = NULL)
plot_admix(sim.X = dat.Y[['data.brute']], sim.Y=NULL, user.bounds=NULL, support="continuous")
## Densite apres avoir rendu la composante connue uniforme:
dat.Y[['data.transform']] <- empiricalCDF.MW(dat.Y[['data.brute']])
plot_admix(sim.X=dat.Y[['data.transform']], sim.Y=NULL, user.bounds=NULL, support="continuous")
## Cannot use 'BVdk' estimation since the known component does not look gaussian (no symmetry).
## The known component does not look like a known distribution, thus all distributions are NULL.
orthoBasis_test_H0(data.X = dat.X[['data.brute']], data.Y = dat.Y[['data.brute']], known.p=NULL,
                   comp.dist = list(NULL,NULL,NULL,NULL), comp.param = list(NULL,NULL,NULL,NULL),
                   known.coef = list(g1 = coefs.voieLactee, g2 = coefs.voieLactee), K = 3,
                   nb.ssEch=2, s = 0.49, var.explicit = FALSE, nb.echBoot=10, support = 'Real',
                   bounds.supp = NULL, est.method = 'PS',
                   uniformized.knownComp_data = list(dat.X[['data.transform']],
                                                     dat.Y[['data.transform']]))
## Try to use 'BVdk' estimator considering the strong following assumption:
list.comp <- list(f1 = NULL, g1 = 'norm',
                  f2 = NULL, g2 = 'norm')
list.param <- list(f1 = NULL, g1 = list(mean = moy.voieLactee, sd = sqrt(var.voieLactee)),
                   f2 = NULL, g2 = list(mean = moy.voieLactee, sd = sqrt(var.voieLactee)))
orthoBasis_test_H0(data.X = dat.X[['data.brute']], data.Y = dat.Y[['data.brute']], known.p=NULL,
                   comp.dist = list.comp, comp.param = list.param,
                   known.coef = list(g1=coefs.voieLactee, g2=coefs.voieLactee), K=3, nb.ssEch=2,
                   s = 0.49, var.explicit = TRUE, nb.echBoot = NULL, support = 'Real',
                   bounds.supp = NULL, est.method = 'BVdk', uniformized.knownComp_data = NULL)
}

}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
