% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_model.R
\name{plot.admix_model}
\alias{plot.admix_model}
\title{Plot method for objects of class \code{admix_model}}
\usage{
\method{plot}{admix_model}(x, n = 1000, main = "Known component distribution", ...)
}
\arguments{
\item{x}{An object of class \code{admix_model}.}

\item{n}{The number of 'x' values to consider for plotting the pdf in the continuous case.}

\item{main}{The title of the plot.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Plots the probability density function of the known component of the admixture model.
}
\examples{
plot(admix_model(knownComp_dist = "norm", knownComp_param = list("mean"=0, "sd"=1)))
plot(admix_model(knownComp_dist = "pois", knownComp_param = list("lambda"=1.5)))

}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
