% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_PS.R
\name{plot.estim_PS}
\alias{plot.estim_PS}
\title{Plots}
\usage{
\method{plot}{estim_PS}(x, ...)
}
\arguments{
\item{x}{An object of class 'estim_PS'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plots
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
