% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatraSen_est_mix_model.R
\name{plot.PS_dist_fun}
\alias{plot.PS_dist_fun}
\title{Results of the computation of the contrast using Patra and Sen approach}
\usage{
\method{plot}{PS_dist_fun}(x, ...)
}
\arguments{
\item{x}{An object of class 'PS_dist_fun'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plots all the attributes of objects of class 'PS_dist_fun'.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
