% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_model.R
\name{print.admix_model}
\alias{print.admix_model}
\title{Print method for objects of class \code{admix_model}}
\usage{
\method{print}{admix_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{admix_model}.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print the information about the distribution of the known component of an admixture model,
as well as the known parameters for this distribution.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
