% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_test.R
\name{print.admix_test}
\alias{print.admix_test}
\title{Print method for objects 'admix_test'}
\usage{
\method{print}{admix_test}(x, ...)
}
\arguments{
\item{x}{An object of class 'admix_test'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print method for objects 'admix_test'
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
