% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_BVdk.R
\name{print.estim_BVdk}
\alias{print.estim_BVdk}
\title{Print method for objects 'estim_BVdk'}
\usage{
\method{print}{estim_BVdk}(x, ...)
}
\arguments{
\item{x}{An object of class 'estim_BVdk'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print an object of class 'estim_BVdk'. An admixture model has probability density function (pdf) l_i such that:
l_i = p_i * f_i + (1-p_i) * g_i, with g_i the known component density.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
