% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_IBM.R
\name{print.estim_IBM}
\alias{print.estim_IBM}
\title{Print method for objects 'estim_IBM'}
\usage{
\method{print}{estim_IBM}(x, ...)
}
\arguments{
\item{x}{An object of class 'estim_IBM'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print an object of class 'estim_IBM'. An admixture model has probability density function (pdf) l_i such that:
l_i = p_i * f_i + (1-p_i) * g_i, with g_i the known component density.
The unknown quantities p_i (i=1,2) are thus listed.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
