% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianity_test.R
\name{print.gaussianity_test}
\alias{print.gaussianity_test}
\title{Print method for objects 'gaussianity_test'}
\usage{
\method{print}{gaussianity_test}(x, ...)
}
\arguments{
\item{x}{An object of class 'gaussianity_test'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print method for objects 'gaussianity_test'
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
