% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM_k_samples_test.R
\name{print.IBM_test}
\alias{print.IBM_test}
\title{Print method for objects 'IBM_test'}
\usage{
\method{print}{IBM_test}(x, ...)
}
\arguments{
\item{x}{An object of class 'IBM_test'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print method for objects 'IBM_test'
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
