% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoComp_mixt.R
\name{print.twoComp_mixt}
\alias{print.twoComp_mixt}
\title{Print method for objects 'twoComp_mixt'}
\usage{
\method{print}{twoComp_mixt}(x, ...)
}
\arguments{
\item{x}{An object of class 'twoComp_mixt'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print an object of class 'twoComp_mixt'. A two-component mixture model has probability density function (pdf) l such that:
l = p * f + (1-p) * g,
where p is the mixing proportion, and f and g are the component distributions.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
