% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{reject_nullHyp.htest}
\alias{reject_nullHyp.htest}
\title{Extractor for object of class 'htest'}
\usage{
\method{reject_nullHyp}{htest}(x)
}
\arguments{
\item{x}{An object of class 'htest'.}
}
\description{
Provide the decision of the statistical test: reject or
do not reject the null hypothesis.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
