% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_cluster.R
\name{summary.admix_cluster}
\alias{summary.admix_cluster}
\title{Results of clustering performed over K populations}
\usage{
\method{summary}{admix_cluster}(object, ...)
}
\arguments{
\item{object}{An object of class 'admix_cluster' (see ?admix_clustering).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the detected clusters among the populations under study, where populations belonging to the same cluster share
the same unknown component distribution. This method also prints the number of clusters, the p-values of statistical
tests performed when building the clusters, the estimated weights of the unknown component distributions inside each
cluster, and the discrepancy matrix. The latter represents some kind of distance between the populations.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
