% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_model.R
\name{summary.admix_model}
\alias{summary.admix_model}
\title{Summary method for objects of class \code{admix_model}}
\usage{
\method{summary}{admix_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{admix_model}.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Summarizes the information related to the known component of the two-component mixture.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
