% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontaminated_dist.R
\name{summary.decontaminated_density}
\alias{summary.decontaminated_density}
\title{Summary method for object of class \code{decontaminated_density}}
\usage{
\method{summary}{decontaminated_density}(object, ...)
}
\arguments{
\item{object}{An object of class \code{decontaminated_density} (see ?decontaminated_density).}

\item{...}{Arguments to be passed to generic method \code{summary}.}
}
\value{
Classical statistical indicators about the decontaminated density.
}
\description{
Summarizes information about the estimated decontaminated density function.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
