% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_BVdk.R
\name{summary.estim_BVdk}
\alias{summary.estim_BVdk}
\title{Summary method for objects 'estim_BVdk'}
\usage{
\method{summary}{estim_BVdk}(object, show.call = TRUE, ...)
}
\arguments{
\item{object}{An object of class 'estim_BVdk'.}

\item{show.call}{A boolean to print the call.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Summarizes the results stored in an object of class 'estim_BVdk'.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
\keyword{internal}
