% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianity_test.R
\name{summary.gaussianity_test}
\alias{summary.gaussianity_test}
\title{Summary method for objects 'gaussianity_test'}
\usage{
\method{summary}{gaussianity_test}(object, ...)
}
\arguments{
\item{object}{An object of class 'gaussianity_test'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Summary method for objects 'gaussianity_test'
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
