% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM_k_samples_test.R
\name{summary.IBM_test}
\alias{summary.IBM_test}
\title{Summary method for objects 'IBM_test'}
\usage{
\method{summary}{IBM_test}(object, ...)
}
\arguments{
\item{object}{An object of class 'IBM_test'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Summary method for objects 'IBM_test'
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
