% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthobasis_test.R
\name{summary.orthobasis_test}
\alias{summary.orthobasis_test}
\title{Summary method for objects of class 'orthobasis_test'}
\usage{
\method{summary}{orthobasis_test}(object, ...)
}
\arguments{
\item{object}{An object of class 'orthobasis_test'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Summary method for objects of class 'orthobasis_test'
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
