% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoComp_mixt.R
\name{summary.twoComp_mixt}
\alias{summary.twoComp_mixt}
\title{Summary method for objects 'twoComp_mixt'}
\usage{
\method{summary}{twoComp_mixt}(object, ...)
}
\arguments{
\item{object}{An object of class 'twoComp_mixt'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Provides statistical indicators of an object of class 'twoComp_mixt'.
A two-component mixture model has probability density function (pdf) l such that:
l = p * f + (1-p) * g,
where p is the mixing proportion, and f and g are the component distributions.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
