% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoComp_mixt.R
\name{twoComp_mixt}
\alias{twoComp_mixt}
\title{Simulation of a two-component mixture model}
\usage{
twoComp_mixt(
  n = 1000,
  weight = 0.5,
  comp.dist = list("norm", "norm"),
  comp.param = list(list(mean = 0, sd = 1), list(mean = 2, sd = 1))
)
}
\arguments{
\item{n}{Number of observations to be simulated.}

\item{weight}{Weight of the first component distribution (distribution f) in the mixture.}

\item{comp.dist}{A list of two elements corresponding to the component distributions (with available names listed in object 'Distribution.df'
in package EnvStats) involved in the mixture model. These elements respectively refer to the two component distributions f and g.
By convention, in the framework of admixture models where one of the two components is unknown, the first element of the list
corresponds to the 'unknown' component distribution, whereas the second one refers to the known one.}

\item{comp.param}{A list of two elements corresponding to the parameters of the component distributions, each element being a list
itself. The names used in each list must correspond to the available parameters listed in object 'Distribution.df'
in package EnvStats. These elements respectively refer to the parameters of f and g distributions of the mixture model.
By convention, in the framework of admixture models where one of the two components is unknown, the first element of the list
corresponds to the 'unknown' component parameters, whereas the second one refers to the known ones.}
}
\value{
An object of class \link[admix]{twoComp_mixt}, containing eight attributes: 1) the number of simulated observations, 2) the simulated mixture
data, 3) the support of the distributions, 4) the name of the component distributions, 5) the name of the parameters of the
component distributions and their values, 6) the mixing proportion, 7) the observations coming from the first component,
8) the observations coming from the second component.
}
\description{
Simulate a two-component mixture model following the probability density function (pdf) \eqn{\ell} such that
\deqn{
  \ell = p f + (1 - p) g,
}
with \eqn{f} and \eqn{g} the mixture component distributions, and \eqn{p} the mixing weight.
}
\examples{
## Mixture of continuous random variables:
sim.X <- twoComp_mixt(n = 1200, weight = 0.7,
                      comp.dist = list("norm", "exp"),
                      comp.param = list(list("mean"=-3, "sd"=0.5),
                                        list("rate"=1)))
print(sim.X)
data.X <- get_mixture_data(sim.X)
plot(density(data.X))

## Mixture of discrete random variables:
sim.Y <- twoComp_mixt(n = 1800, weight = 0.7,
                      comp.dist = list("multinom", "multinom"),
                      comp.param = list(list("size"=1, "prob"=c(0.3,0.4,0.3)),
                                        list("size"=1, "prob"=c(0.6,0.2,0.2))))
print(sim.Y)

}
\seealso{
\code{\link[=get_mixture_data]{get_mixture_data()}} to access the simulated mixture data.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
