% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_distribution}
\alias{validate_distribution}
\title{Check the validity of the specified distribution and parameter(s)}
\usage{
validate_distribution(dist, params)
}
\arguments{
\item{dist}{A single string naming the distribution under consideration.}

\item{params}{A character vector composed of the names of the parameters for a given distribution.}
}
\value{
A vector composed of the names of the parameters of the distribution.
}
\description{
Check the validity of the specified distribution and parameter(s)
}
\examples{
validate_distribution(dist = "norm", params = c("mean" = 0, "sd" = 1))

}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
