% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{which_rank.htest}
\alias{which_rank.htest}
\title{Extractor for object of class 'htest'}
\usage{
\method{which_rank}{htest}(x)
}
\arguments{
\item{x}{An object of class 'htest'.}
}
\description{
Provide the selected rank of the test statistic (connected to the expansion order
of the densities in the orthonormal polynomial basis if method 'poly' was chosen;
or to the number of terms, i.e. discrepancies between couples of samples, included
in the test statistic).
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
