% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_discrepancy_id}
\alias{get_discrepancy_id}
\title{Extractor for object of class 'IBM_test'}
\usage{
get_discrepancy_id(x)
}
\arguments{
\item{x}{An object of class 'IBM_test'.}
}
\description{
Provide the matrix storing the identifiers of discrepancies using Inversion-Best Matching
approach between all couples among the K samples under study.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
