% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_cluster.R
\name{get_discrepancy_matrix}
\alias{get_discrepancy_matrix}
\title{Extractor for object of class 'admix_cluster'}
\usage{
get_discrepancy_matrix(x)
}
\arguments{
\item{x}{An object of class 'admix_cluster'.}
}
\description{
Provide the matrix storing discrepancies between all couples
among the K samples under study.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
