% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_discrepancy_rank}
\alias{get_discrepancy_rank}
\title{Extractor for pairwise discrepancy rankings}
\usage{
get_discrepancy_rank(x)
}
\arguments{
\item{x}{An object of class \code{IBM_test}.}
}
\value{
A matrix of ranks, from the closest couple (rank 1) in terms of discrepancy
measure to the most different one.
}
\description{
Provide the matrix storing the ranks of discrepancies using Inversion-Best Matching
approach between all couples among the K (K>2) samples under study.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
