% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_known_component}
\alias{get_known_component}
\title{Extractor for objects of class 'admix_estim', 'htest' or 'admix_cluster'}
\usage{
get_known_component(x)
}
\arguments{
\item{x}{An object of class 'admix_estim', 'htest', or 'admix_cluster'.}
}
\description{
Get the known component(s) of admixture model(s) considered for estimation, test, or clustering.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
