% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoComp_mixt.R
\name{get_mixture_data}
\alias{get_mixture_data}
\title{Extractor for object of class 'twoComp_mixt'}
\usage{
get_mixture_data(x)
}
\arguments{
\item{x}{An object of class 'twoComp_mixt'.}
}
\description{
Get the mixture data generated from method 'twoComp_mixt'.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
