% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milkyWay.R
\docType{data}
\name{milkyWay}
\alias{milkyWay}
\title{Heliocentric velocity measured for the Milky Way (from Walker, M. G., M. Mateo, E. W. Olszewski, O. Y. Gnedin, X. Wang,
B. Sen, and M. Woodroofe (2007). Velocity dispersion profiles of seven dwarf spheroidal galaxies. Astrophysical J. 667(1), L53–L56).}
\format{
A data frame with 170,601 rows and 1 column:
\describe{
\item{V1}{Heliocentric velocity measurements of the Milky way}
}
}
\source{
\url{https://www.aanda.org/articles/aa/full_html/2018/08/aa32905-18/aa32905-18.html}
}
\usage{
milkyWay
}
\description{
Heliocentric velocity measured for the Milky Way (from Walker, M. G., M. Mateo, E. W. Olszewski, O. Y. Gnedin, X. Wang,
B. Sen, and M. Woodroofe (2007). Velocity dispersion profiles of seven dwarf spheroidal galaxies. Astrophysical J. 667(1), L53–L56).
}
\keyword{datasets}
