% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoComp_mixt.R
\name{plot.twoComp_mixt}
\alias{plot.twoComp_mixt}
\title{Plots several mixture densities on the same graph}
\usage{
\method{plot}{twoComp_mixt}(x, add.plot = FALSE, ...)
}
\arguments{
\item{x}{Object of class 'twoComp_mixt' from which the density will be plotted.}

\item{add.plot}{(default to FALSE) Option to plot another mixture distribution on the same graph.}

\item{...}{further classical arguments and graphical parameters for methods plot and hist.}
}
\value{
a plot with the densities of the samples provided as inputs.
}
\description{
Plots the empirical densities of the samples with optional arguments to improve the visualization.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
