% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_cluster.R
\name{print.admix_cluster}
\alias{print.admix_cluster}
\title{Print method for object of class 'admix_cluster'}
\usage{
\method{print}{admix_cluster}(x, ...)
}
\arguments{
\item{x}{An object of class 'admix_cluster' (see ?admix_clustering).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print the main results when clustering the unknown component distributions coming from various
admixture samples, i.e. the obtained clusters.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
