% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontaminated_dist.R
\name{print.decontaminated_density}
\alias{print.decontaminated_density}
\title{Print method for object of class 'decontaminated_density'}
\usage{
\method{print}{decontaminated_density}(x, ...)
}
\arguments{
\item{x}{An object of class 'decontaminated_density' (see ?decontaminated_density).}

\item{...}{Arguments to be passed to generic method 'plot', such as graphical parameters (see par).}
}
\value{
The function related to the estimated decontaminated density.
}
\description{
Print some overview of the decontaminated density function.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
