% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_BVdk.R
\name{print.estim_BVdk}
\alias{print.estim_BVdk}
\title{Print method for objects 'estim_BVdk'}
\usage{
\method{print}{estim_BVdk}(x, ...)
}
\arguments{
\item{x}{An object of class 'estim_BVdk'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print the results stored in an object of class 'estim_BVdk'.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
\keyword{internal}
