% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_IBM.R
\name{print.estim_IBM}
\alias{print.estim_IBM}
\title{Print method for objects of class 'estim_IBM'}
\usage{
\method{print}{estim_IBM}(x, ...)
}
\arguments{
\item{x}{An object of class 'estim_IBM'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Print the results stored in an object of class 'estim_IBM'.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
\keyword{internal}
