% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_cluster.R
\name{summary.admix_cluster}
\alias{summary.admix_cluster}
\title{Summary method for object of class 'admix_cluster'}
\usage{
\method{summary}{admix_cluster}(object, ...)
}
\arguments{
\item{object}{An object of class 'admix_cluster' (see ?admix_clustering).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarizes the results obtained when clustering the unknown component distributions coming from various
admixture samples.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
