% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_model.R
\name{summary.admix_model}
\alias{summary.admix_model}
\title{Summary method for objects of class 'admix_model'}
\usage{
\method{summary}{admix_model}(object, ...)
}
\arguments{
\item{object}{An object of class 'admix_model'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Summarizes the information related to the known component of the two-component mixture.
An admixture model has probability density function (pdf) l_i such that:
l_i = p_i * f_i + (1-p_i) * g_i, with g_i the known component density.
The unknown quantities are therefore p_i and f_i.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
