% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_IBM.R
\name{summary.estim_IBM}
\alias{summary.estim_IBM}
\title{Summary method for objects 'estim_IBM'}
\usage{
\method{summary}{estim_IBM}(object, ...)
}
\arguments{
\item{object}{An object of class 'estim_IBM'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Summarizes the results stored in an object of class 'estim_IBM'.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
