% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_PS.R
\name{summary.estim_PS}
\alias{summary.estim_PS}
\title{Summary method for objects 'estim_PS'}
\usage{
\method{summary}{estim_PS}(object, ...)
}
\arguments{
\item{object}{An object of class 'estim_PS'.}

\item{...}{A list of additional parameters belonging to the default method.}
}
\description{
Summarizes the results stored in an object of class 'estim_PS'.
}
\author{
Xavier Milhaud \href{mailto:xavier.milhaud.research@gmail.com}{xavier.milhaud.research@gmail.com}
}
